For #673: add AbstractBaseSQLParsingEngineTest
